backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/LLM
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.5
pinecone-client==2.2.4
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# AWS
boto3==1.34.0
aioboto3==12.3.0

# Utilities
httpx==0.25.2
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 